Uses GCD method instead of dividing 1000 primes.
Replaces the previous sieve method of using a list of the first 1000 primes
and taking the modulus with the primality candidate.

The new method uses the product of the first 1000 primes to compute the GCD of
the candidate with the product. If the GCD is not one, then the candidate can
be confirmed as composite.